refactor: export readDat from compile package
