+++ packages/compile/src/index.js
@@ -1,3 +1,4 @@
 // Copyright (c) 2022 Climate Interactive / New Venture Fund
 
 export { createParser, parseModel } from './Parser.js'
+export { readDat } from './ReadDat.js'
